my_Projects/Test_Planner_Agent/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
anthropic>=0.25.0
groq>=0.9.0
markdown>=3.6
fpdf2>=2.7.9
# weasyprint>=61.0   # Uncomment if GTK is available on your system (better PDF quality)
